agents_catalog/32-sila2-lab-automation-agent-backup-20251121-064659/requirements-minimal.txt
boto3
requests==2.31.0
grpcio>=1.50.0
protobuf>=4.21.0
grpcio-tools>=1.50.0